Fix `#include`s in `src/wallet`
